code/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
et-xmlfile==1.1.0
executing==1.2.0
fastdtw==0.3.4
fastjsonschema==2.16.3
fonttools==4.39.3
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
ibm-cloud-sdk-core==3.16.5
ibm-platform-services==0.34.2
idna==3.4
imbalanced-learn==0.10.1
ipykernel==6.22.0
ipython==8.23.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.25
jsonpointer==2.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.5
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.1.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.26.0
kiwisolver==1.4.4
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==7.1.2
notebook_shim==0.2.3
ntlm-auth==1.5.0
numba==0.57.0
numpy==1.23.5
openpyxl==3.1.2
openTSNE==1.0.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.43
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pylatexenc==2.10
pynndescent==0.5.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qiskit==0.42.1
qiskit-aer==0.12.0
qiskit-ibm-provider==0.5.2
qiskit-ibm-runtime==0.9.3
qiskit-ibmq-provider==0.20.2
qiskit-machine-learning==0.6.0
qiskit-terra==0.23.3
qtconsole==5.4.2
QtPy==2.3.1
referencing==0.34.0
requests==2.31.0
requests-ntlm==1.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rustworkx==0.12.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.14.2
stevedore==5.0.0
symengine==0.10.0
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.4
tqdm==4.65.0
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2023.3
umap-learn==0.5.3
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
widgetsnbextension==4.0.7